langchain-kredo/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "langchain-kredo"
version = "0.3.0"
description = "LangChain integration for the Kredo agent attestation protocol"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Jim Motes" },
    { name = "Vanguard" },
]
keywords = ["langchain", "kredo", "attestation", "trust", "agents", "ed25519"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "langchain-core>=0.3.0,<1.0.0",
    "kredo>=0.4.0",
    "pynacl>=1.5.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.urls]
Homepage = "https://aikredo.com"
Repository = "https://github.com/jimmotes2024/kredo"
Documentation = "https://aikredo.com"

[tool.hatch.build.targets.wheel]
packages = ["langchain_kredo"]
